% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuildDatasets.R
\name{BuildAllDatasets}
\alias{BuildAllDatasets}
\title{Builds all datasets of DAPARdata}
\usage{
BuildAllDatasets()
}
\value{
Datasets
}
\description{
Builds all datasets of DAPARdata
}
\examples{
BuildAllDatasets()

}
