% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{DAPARdata}
\alias{DAPARdata}
\title{Available datasets in the DAPARdata package}
\usage{
DAPARdata()
}
\value{
A list of datasets
}
\description{
This function lists the datasets available in DAPARdata package
}
\examples{
DAPARdata()
}
