% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{Exp2_R100_pept}
\alias{Exp2_R100_pept}
\title{Exp2_R100_pept dataset}
\format{An object of class \code{\link{MSnSet}} related to peptides
quantification. It contains 6 samples divided into two conditions
(1 fmol and 100 fmol) and 5684 peptides.

The data frame exprs(Exp2_R100_pept) contains six columns that are the 
quantitation of peptides for the six replicates. 

The data frame fData(Exp2_R100_pept) contains the meta data about the peptides

The data frame pData(Exp2_R100_pept) contains the experimental design and gives 
few informations about the samples.}
\usage{
data(Exp2_R100_pept)
}
\value{
An object of class \code{\link{MSnSet}} related to peptides
quantification.
}
\description{
This dataset is xxxxxxx

The dataset is either available as a CSV file 
(see inst/extdata/Exp2_R100_pept.txt), or as a \code{\link{MSnSet}} structure 
(Exp2_R100_peptt.MSnset). In the latter case, the quantitative data are those of 
the raw intensities.
}
\references{
xxxxxxx
}
\keyword{data}
\keyword{datasets}

