% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{Exp2_R2_prot}
\alias{Exp2_R2_prot}
\title{Exp2_R2_prot dataset}
\format{An object of class \code{\link{MSnSet}} related to proteins
quantification. It contains 6 samples divided into two conditions
(25fmol and 50fmol) and 948 proteins.

The data frame exprs(Exp2_R2_prot) contains six columns that are the 
quantitation of proteins for the six replicates. 

The data frame fData(Exp2_R2_prot) contains the meta data about the proteins.

The data frame pData(Exp2_R2_prot) contains the experimental design and gives 
few informations about the samples.}
\usage{
data(Exp2_R2_prot)
}
\value{
An object of class \code{\link{MSnSet}} related to proteins
quantification.
}
\description{
This dataset is xxxxxxx

From a statistical viewpoint, the goal is to find which proteins are 
differentially abundant between the two conditions among the 948 
quantified proteins. Ideally, the 46 quantified human proteins (out of the 
original 48 ones) should be concluded as differentially abundant.

The dataset is either available as a CSV file 
(see inst/extdata/Exp2_R2_prot.txt), or as a \code{\link{MSnSet}} structure 
(Exp2_R2_prot.MSnset). In the latter case, the quantitative data are those of 
the raw intensities.
}
\references{
xxxxxxx
}
\keyword{data}
\keyword{datasets}

