% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuildDatasets.R
\name{Set_X_CC_infos}
\alias{Set_X_CC_infos}
\title{Adjacency matrix and connected components information}
\usage{
Set_X_CC_infos(obj)
}
\arguments{
\item{obj}{An object of class `MSnSet`}
}
\value{
An object of class `MSnSet`
}
\description{
Adjacency matrix and connected components information
}
\examples{
Set_X_CC_infos(Exp1_R25_pept)

}
