% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{UPSprot25}
\alias{UPSprot25}
\title{UPSprot25 dataset}
\format{An object of class \code{\link{MSnSet}} related to proteins
quantification. It contains 6 samples divided into two conditions
(25fmol and 10fmol) and 2384 proteins.

The data frame exprs(UPSprot25) contains six columns that are the 
quantitation of proteins for the six replicates. 

The data frame fData(UPSprot25) contains the meta data about the proteins.

The data frame pData(UPSprot25) contains the experimental design and gives 
few informations about the samples.}
\usage{
data(UPSprot25)
}
\value{
An object of class \code{\link{MSnSet}} related to proteins
quantification.
}
\description{
This dataset is the final outcome of a quantitative mass 
spectrometry-based proteomic analysis of two samples containing different 
concentrations of 48 human proteins (UPS1 standard from Sigma-Aldrich) 
within a constant yeast background (see Giai Gianetto et al. (2016) for 
details). It contains the abundance values of the different human and 
yeast proteins identified and quantified in these two conditions. The two 
conditions represent the measured abundances of proteins when respectively 
25fmol and 10fmol of UPS1 human proteins were mixed with the yeast extract 
before mass spectrometry analyses. Three technical replicates were acquired 
for each condition.

To identify and quantify proteins, spectra were searched using MaxQuant 
(version 1.5.1.2) against the Uniprot database, the UPS database and the 
frequently observed contaminants database. Maximum false discovery rates 
were set to 0.01 at peptide and protein levels by employing a reverse 
database strategy. The abundance values of the dataset were obtained from 
LFQ values calculated using MaxQuant from MS intensity of unique peptides 
(see Cox et al. (2014)).

From a statistical viewpoint, the goal is to find which proteins are 
differentially abundant between the two conditions among the 2384 
quantified proteins. Ideally, the 46 quantified human proteins (out of the 
original 48 ones) should be concluded as differentially abundant.

The dataset is either available as a CSV file 
(see inst/extdata/UPSprot25.txt), or as a \code{\link{MSnSet}} structure 
(UPSprot25.MSnset). In the latter case, the quantitative data are those of 
the raw intensities.
}
\references{
Cox J., Hein M.Y., Luber C.A., Paron I., Nagaraj N., Mann M. 
Accurate proteome-wide label-free quantification by delayed normalization 
and maximal peptide ratio extraction, termed MaxLFQ. Mol Cell Proteomics. 
2014 Sep, 13(9):2513-26.

Giai Gianetto, Q., Combes, F., Ramus, C., Bruley, C., Coute, Y., Burger, 
T. (2016). Calibration plot for proteomics: A graphical tool to visually 
check the assumptions underlying FDR control in quantitative experiments. 
Proteomics, 16(1), 29-32.
}
\keyword{data}
\keyword{datasets}

